backend/requirement.txt
Flask==2.2.5
flask-socketio==5.3.3
eventlet==0.33.3
azure-messaging-webpubsubservice==1.2.1
